M1 phase C closeout — audit fixes (2 bugs + spill-ref front door)
Three adversarial audits: xlsx/format/honesty 8/8 verified zero
refutations; correctness audit found 2 real bugs + 1 reachability gap,
all fixed with the regressions the audit flagged as missing:
- DATEDIF(...,"MD") u32 underflow -> 4.29e9 garbage when the start day
  exceeds the preceding month length; now i64, yields Excel's -2
  bug-for-bug (date2.rs).
- FACTDOUBLE(-1) returned #NUM!; ECMA-376 domain is n>=-1 (empty
  product = 1); guard fixed to n<-1, -2 still #NUM! (math2.rs).
- A1# spill-ref was unreachable from typed input (# routed
  unconditionally to error-literal lexing). # is now context-sensitive
  (#+letter = error literal, else the postfix spill operator);
  Expr::SpillRef wired in the Pratt parser (the AST/eval/print support
  already existed). SUM(A1#)=45 verified e2e.

994 cargo tests, 94 vitest, clippy+fmt clean, coverage gate 286/0/5,
wasm 933 KB. Everything else (fin/stat/lookup2/math2/text2/spill/
structured-refs/pagination) audit-confirmed solid.

+++ sheet-fn/src/families/date2.rs
@@ -277,7 +277,12 @@ pub fn datedif(args: &[Arg], ctx: &crate::ctx::EvalCtx) -> CellValue {
                 // Days in the month before the end month (where the borrow
                 // comes from).
                 let (py, pm) = if em == 1 { (ey - 1, 12) } else { (ey, em - 1) };
-                (days_in_month(py, pm) + ed - sd) as i64
+                // Cast to i64 BEFORE subtracting: all three are u32, and
+                // `days_in_month(prev) + ed` can be < sd (start day past the
+                // preceding month's length, e.g. Jan-31 -> early March), which
+                // would underflow in u32. The negative result is Excel's own
+                // documented "MD" bug-for-bug behaviour (-2 here), not a panic.
+                days_in_month(py, pm) as i64 + ed as i64 - sd as i64
             }
         }
         "YM" => {
